fix Get balance

+++ internal/adapters/database/postgres/tables.go
@@ -11,7 +11,7 @@ const (
 		    id SERIAL PRIMARY KEY,
 		    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
 		    order_id VARCHAR(255) UNIQUE NOT NULL,
-		    accrual INTEGER NOT NULL,
+		    accrual FLOAT NOT NULL,
 			status VARCHAR(255) NOT NULL,
 		    uploaded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 		);
@@ -21,7 +21,7 @@ const (
 		    order_id VARCHAR(255),
 		    processed_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 		    action VARCHAR(32) NOT NULL,
-		    amount INTEGER NOT NULL CHECK (amount >= 0)
+		    amount FLOAT NOT NULL CHECK (amount >= 0)
 		);
 		CREATE INDEX IF NOT EXISTS order_id_index ON balances(order_id);
 		DO
